[refactor] UserImage 패키지 이동
