+++ src/main/java/com/prgrms/ohouse/domain/user/model/UserImage.java
@@ -1,10 +1,11 @@
-package com.prgrms.ohouse.domain.common.file;
+package com.prgrms.ohouse.domain.user.model;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 
+import com.prgrms.ohouse.domain.common.file.StoredFile;
 import com.prgrms.ohouse.domain.user.model.User;
 
 import lombok.AccessLevel;
@@ -12,7 +13,7 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserImage extends StoredFile{
+public class UserImage extends StoredFile {
 
 	public UserImage(String originalFileName, String url, User user) {
 		super(originalFileName, url);
